Core multi-page column index support (#138)
* Generate mutiple pages in column index test

In order to exercise all behavirs of the column index, it's necessary to
generate a parquet file with multiple pages. The existing test is
modified to create a file with a random buffer size between 1 and 50.

* Update tests for column index direction

Test that the column index is reporting the correct order by observing
all values in the index, maintaining a list of min and max values per
page, and comparing the observed order to the order reported.

* Statically order buffer column indexes

A buffer column index will only have a single page so its order doesn't
matter. Hard code to always return ascending order to match the behavior
of fileColumnIndex with a single item min/max index.

* Fix bytes ordering

The bytes order test accidentally created inputs containing all equal
values so the ordering implementation was never exercised. Fixing the
test exposed a bug in the order implementation which is fixed here.

* Single item slices unordered

The order of a single item slice is irrelevant. Update ordering logic to
report them as unordered so that single-page column indexes report
unordered.

* Clarify ordered test helper name

* Clarify order test helper

Update order test helper to work with the `parquet.Type.Compare` API.
Refactor how `Compare()` is called to be less surprising.