Refactor Conversion Code to be schema-aware (#319)
convert: introduce schema-aware conversions

Prior to this changeset, conversions between rows having different
schemas was not schema aware.  The strategy prior to this PR was simply
to append all the columns sequentially, causing issues when trying to
parse the new row during reconstruction.

Conversion now uses the schema to determine how to build the row, in a
manner which is consistent with what reconstruct expects.