schema: handle interfaces like pointers
When a schema is defined, and we're instantiating a generic buffer using
`any`, we need to dereference the `any` (`interface{}`) type to get the
underlying element to be written (which should also match the schema
given).  Currently there's no verification that they match to my
knowledge but it maybe a good idea to add an additional check somehow.