buffer_go18: enforce schema presence at instantiation
We could infer the schema from the first element in a slice lazily in
some cases, but error handling in subsequent calls to functions that
depend on the schema may be difficult in the cases where the generic
buffer was instantiated with an interface type parameter.

If there's a compelling use case for that, we may add the functionality
down the road.