search: add comments to search.go (#388)
* search: add comments to search.go

I was going through this function to check my understanding of page
indexes while thinking about granules in TraceDB.

I believe that granules in other databases like clickhouse and frostdb
are the equivlant of column pages in parquet.  The page index is the
sparse index over the pages (granules) wherein the data being searched
for is contained.

This check on the binary search mostly confirms that suspicion, if I'm
not mistaken.  Please correct me if I'm misunderstaneding something.

Also, as part of this, I found what is possibly an optimization to avoid
skipping ahead a page, by comparing the current value to the max in the
current page, instead of letting the binary search check the next page,
then falling through to the final if-block.

It may or may not be a bug also, if, for example, the following happens:

you have pages:

[1,2,3]
[3,4,5]
[6,7,8]

Searching for 3 in this case would not return the first page and instead
return the second, missing values of 3 in the first page.

I'll have to look at writing a test for this, but the current search
test will pass creating this scenario without altering the test
somewhat (since it's only looking for presence, not the expected index)

* clarify, and handle case properly, I think

* nit

* search: remove nextIdx-1 > curIdx bounds check, and MaxValue check

The bounds check actually misses some cases where we do want to assign
topIdx = nextIdx

e.g.
searchValue:40
pages:
[1,2,3]
[10,20,30]
[30,40,40]
[40,50,60]

And the MaxValue check is unecessary since we would want to search
pages[curIdx] in the case where minValue == value (previously handled by
the default case statement, which is no longer necessary removing the
MaxValue check.)

* failing test for #388