Convert values between Parquet logical types (#393)
To support deserializing Parquet timestamp values for #266, there needs
to be a way to convert between different Parquet types. For example,
there might be a TIMESTAMP(isAdjustedToUTC=true, unit=MILLIS) value
stored in a file and it needs to be deserialized into a time.Time which
expects a number of nanoseconds. This PR adds a
ConvertValue(val Value, typ Type) function to the parquet.Type interface
which accepts the source Value and Type and return a Value converted to
the Type of the receiver.