writer,bloom: make filter bits per value configurable
Previously we would use 10 bits per value for all bloom filters.
However, in our quest for more performance, we may try to tweak these
values as appropriate for the dataset being optimized.

Add a bitsPerValue argument to SplitBlockFilter so that callers are
required to pass this value, and remove bitsPerValue from the
BloomFilterColumn interface.

To update existing code to work with the new API with no change in
functionality, update:

    SplitBlockFilter("first_name")

to:

    SplitBlockFilter(10, "first_name")