Reuse reconstruction columns slice to reduce allocations
Create an internal method on Schema to reconstruct a value, passing in
a [][]Value to use as columns. Use this internal method, rather than
Reconstruct, when reading rows in GenericReader.

Calling Reconstruct on every row being read, and constructing a new
[][]Value, in aggregate, accounts for the majority of allocations while
reading parquet files and induces an unnecessarily large GC overhead.